+++ backend/migrations/20210617020426-create-user-group.js
@@ -23,6 +23,8 @@ module.exports = {
           key: "id",
           as: "GroupId",
         },
+        onUpdate: 'cascade',
+        onDelete: 'cascade'
       },
       createdAt: {
         allowNull: false,